Better code structure for utils
