+++ rackio/utils/auth.py
@@ -1,0 +1,22 @@
+# -*- coding: utf-8 -*-
+"""rackio/utils/auth.py
+
+This module implements Authentication Utility Functions.
+"""
+
+from uuid import uuid1
+from hashlib import md5
+
+def generate_key():
+
+    return str(uuid1())
+
+def hash_password(password):
+
+    _hash = md5(password.encode())
+
+    return _hash.hexdigest()
+
+def verify_password(_hash, password):
+
+    return _hash == hash_password(password)
